Julie/requirements.txt
spotipy
SpeechRecognition
pyttsx3
python-dotenv
